Single_agent_using_Groq_app/requirements.txt
# Core Agent Framework
crewai
crewai-tools

# LLM Connectors
langchain-groq
litellm[proxy]
fastapi
uvicorn

# UI and Environment
streamlit
python-dotenv
